block: turn the nr_iovecs argument to bio_alloc* into an unsigned short
mainline inclusion
from mainline-v5.12-rc1
commit 0f2e6ab851ae146c468bc5151c302c6e2473f70a
bugzilla: https://gitee.com/openeuler/kernel/issues/IB7FJU

Reference: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=0f2e6ab851ae146c468bc5151c302c6e2473f70a

--------------------------------

The bi_max_vecs and bi_vcnt fields are defined as unsigned short, so
don't allow passing larger values in.


Conflicts:
  block/bio.c
  include/linux/bio.h
[Commit 3175199ab0ac ("block: split bio_kmalloc from bio_alloc_bioset")
change the comment of bio_alloc_bioset and move bio_kmalloc from
include/linux/bio.h to block/bio.c]

+++ block/bio.c
@@ -419,7 +419,7 @@ static void punt_bios_to_rescuer(struct bio_set *bs)
  *   RETURNS:
  *   Pointer to new bio on success, NULL on failure.
  */
-struct bio *bio_alloc_bioset(gfp_t gfp_mask, unsigned int nr_iovecs,
+struct bio *bio_alloc_bioset(gfp_t gfp_mask, unsigned short nr_iovecs,
 			     struct bio_set *bs)
 {
 	gfp_t saved_gfp = gfp_mask;
